+++ cpukit/score/src/heapsizeofuserarea.c
@@ -47,7 +47,7 @@ bool _Heap_Size_of_alloc_area(
     return false;
   }
 
-  *alloc_size = (uintptr_t) next_block + HEAP_ALLOC_BONUS - alloc_begin;
+  *alloc_size = (uintptr_t) next_block + HEAP_ALLOC_BONUS - (size_t) alloc_begin;
 
   return true;
 }
